Started factoring out routes Relates #21
