+++ src/handler.js
@@ -1,0 +1,30 @@
+var fs = require('fs');
+var path = require('path');
+
+const handlers = {
+  home: (req, res) => {
+    const filePath = path.join(__dirname, "..", "public", "index.html");
+    fs.readFile(filePath, (error, file) => {
+      if (error) {
+        res.writeHead(500, {
+          "Content-type": "text/html"
+        });
+        res.end("<h1>So sorry, we've had a problem on our end.</h1>");
+
+      } else {
+        res.writeHead(200, {
+          "Content-type": "text/html"
+        });
+        res.end(file);
+      }
+    });
+  },
+  search: (req, res) => {
+    res.writeHead(200, {
+      "Content-type": "text/html"
+    });
+    res.end();
+  }
+}
+
+module.exports = handlers;
